+++ cfftlog/cfftlog.h
@@ -5,8 +5,8 @@ typedef struct config {
 	long N_pad;
 } config;
 
-void cfftlog(double *x, double *fx, long N, config *config, int ell, double *y, double *Fy);
+void cfftlog(double *x, double *fx, long N, config *config, double ell, double *y, double *Fy);
 
-void cfftlog_ells(double *x, double *fx, long N, config *config, int* ell, long Nell, double **y, double **Fy);
+void cfftlog_ells(double *x, double *fx, long N, config *config, double* ell, long Nell, double **y, double **Fy);
 
-void cfftlog_ells_increment(double *x, double *fx, long N, config *config, int* ell, long Nell, double **y, double **Fy);
+void cfftlog_ells_increment(double *x, double *fx, long N, config *config, double* ell, long Nell, double **y, double **Fy);
